refactor: split course flow modules
