+++ docs/dev-logs/README.md
@@ -16,6 +16,7 @@
 - 로그는 `왜 바꿨는지`, `무엇을 바꿨는지`, `어떻게 검증했는지`만 짧게 적는다.
 
 ## 최신 로그
+- `2026-04-08-refactor-course-flow-structure.md`
 - `2026-04-08-course-progress-flow.md`
 - `2026-04-07-ai-collaboration-dedup.md`
 - `2026-04-07-ai-collaboration-official-rules.md`
